update build method

+++ engine/src/defines.h
@@ -1,18 +1,22 @@
 #pragma once
 
-typedef unsigned char  u8;
-typedef unsigned short u16;
-typedef unsigned int   u32;
-typedef unsigned long  u64;
+// Unsigned int types.
+typedef unsigned char      u8;
+typedef unsigned short     u16;
+typedef unsigned int       u32;
+typedef unsigned long long u64;
 
+// Signed int types.
 typedef signed char      i8;
 typedef signed short     i16;
 typedef signed int       i32;
 typedef signed long long i64;
 
+// Floating point types
 typedef float  f32;
 typedef double f64;
 
+// Boolean types
 typedef int  b32;
 typedef char b8;
